Adjust Last-Translator as agreed between Mario and me
The following rule was agreed:
If the Last-Translator is Mario, then Mario will take care of updates,
possible after review on debian-l10n-german.

Otherwise Helge will take care of the updates if possible.

Note: This rule is under trial. If it does not work out, it might be
changed/adjusted.