Add Romanian to partially unmaintained languages
For better understanding, when we introduced Romanian as another supported
language, Mihai Cristescu signalized to be willing to maintain the
translations of the nano man manpages only. There was no talk of maintaining
the other files we've imported. So actually we need some translator(s) who
are willing to work on the non-nano files.