Revert "update shebang to force python3.7"
This reverts commit b504d603d2e7f5e8998b39a4f704924c6ee26b5c.

A specific python3 version is not correct, as this would render
the scripts unusable on each system with a different minor
version. For example, the version of python in unstable and
testing is currently python 3.8.2, version 3.7 is no longer
available.

The correct way to fix the bug would be to catch the exception.