[fr] Make URLs to www.blaess.fr secure
This commit only affects comments specific to the french translation.

It replaces http://www.blaess.fr with https://www.blaess.fr

In particular, it does not touch any msgstr so the translations
themselves are not modified by this commit.

Its only purpose is to minimize the number of occurences of non-secure
URLs (http:) in the french translations.