Update all templates
For updating really all templates, I've reanimated the old script
update-all-templates.sh, which is the predecessor of the current
update-all-templates.py. The latter failed in some cases, for example, when
new packages got added to the package lists. It works by updating only the
the templates for which the upstream man pages have been changed, but it
needs a valid JSON files written by /upstream/update-distribution-manpages.sh.
If this files are missing, it refuses to update the templates completely.