Refactor create-authors-file.sh with a for loop
Loop through a list of the supported languages to run
all commands that are currently ran today, without
having to repeat commands, hence avoiding possible
copy-and-paste errors.

Each entry of the language list consists of a language name
surround by double quotes, and its language code, in that
order. Changing the order of code/name or using a
non-double-quoted language name will break the script.

Also, replace sed+uniq with 'sed' -u for the same result,
and change multiple 'echo' calls to a single multi-line
string 'echo' call.

language_list is an array, so shebang had to be changed
from sh to bash because array is not defined on POSIX sh,
dash, etc. (see shellcheck report)