Get rid of the annoying grub-mknetdir error message, second attempt
Hopefully it works... How to determine and then copy and paste a
non-printable character? Kate doesn't see it, Nano detects is as ^K, but in
fact it is a vertical tabulator (U+000b), for whatever reasons. So I hope
that next time we do an upstream update, we don't have this thing anymore,
which slows down the template and .po file scripts significantly.