[pl] Rename the 'binutils-x86-64-linux-gnu*' man pages
Debian ships architecture-specific packages for GNU Binutils. At a first
glance, the packages seem to contain man pages, but in fact, the files are
only links to those from the binutils-common package. I've renamed the
existing .po files to their basenames (ranlib.1.po and so on) to be able
to sync the files with real templates. Moreover, I've removed the
corresponding entries from packages.txt in Debian Bullseye and Unstable.