scripts: Switch license addendum to plain text
We need to special case mdoc syntac and as the addendum of the license
always follows the "translation" addendum, we can move the former to
plain text and merge it (before calling po4a) with the latter.

This way we can maintain two versions: One "ordinary" and one for mdoc.

Currently, the list of files which use mdoc is hardcoded.

Secondly, the implementation only considers German for the moment, once
the testing is completed, the same change needs to be done to other
languages as well.