upstream/debian-bullsey: Include backports for Bullseye
This was the original behaviour and has been requested by the current
maintainer. Even though backports are enabled individually, a user
interested in a certain software *and* using manpages-XX from backports
already might likely also include the backport for his favorite
software.

Of course, this assumption may be wrong, but then usually the backport
is a superset of the original software and thus all of the stable
version is also described, just some "extra" is added.

Security is missing, here the Debian website is broken,
see Debian bug #1003604