New script to create the files "untranslated.txt" automatically.
This script creates the file untranslated.txt for all languages. It
assignes each man page a package, based on the bug entry in upstreambug.
If a man page is coming from several packages (due to several
distributions using different sources), a random asignment is done.

There is a placeholder for an enhancement in the script.