public: New script to generate web view of untranslated files
This script mainly reproduces the previous (but still active) code,
except that it does not try to assign distributions (this does not make
much sense anyhow).

It can (and might be) improved, e.g. providing links or similar.

Note that this script is not yet activated and hence by default does not
produce output.